feat: add custom hook useAlert
